[AArch64]strcmp.S: Improve performance for misaligned strings
Replace the simple byte-wise compare in the misaligned case with a
dword compare with page boundary checks in place.  For simplicity I've
chosen a 4K page boundary so that we don't have to query the actual
page size on the system.

This results in up to 3x improvement in performance in the unaligned
case on falkor and about 2.5x improvement on mustang as measured using
bench-strcmp in glibc.

Change-Id: I0a8513e3d21a0f15fea51635a8e1abc7537812bf
Signed-off-by: spezi77 <spezi7713@gmx.net>